tweedledum Noah, a new beginning.
"And, behold, I, even I, do bring a flood of waters upon the earth, to
destroy all flesh, wherein is the breath of life, from under heaven; and
every thing that is in the earth shall die".

Created an orphan branch because this will be a big change.  I will
reuse a lot of the code but I will change most of the names and use a
terminology closer to what compilers use.