Sketching out the basic infrastructure.
This implements a lexer, a pre-processor and a parser for OpenQASM 2.0
I took most of it from `tweedledee` by taking out the code responsible
for building the AST.  For now the parser just consume the grammar and
instantiate `gate`s.  I also added a tool that reads a `.qasm` file and
dump what comes out of the lexer or the lexer+pre-processor.  It also
has the option to run the parser and will return a boolean indicating
whether the parsing was successful or not---it will accept broken code.
So far is just quite trivial stuff.