Big bang! Add most of the infrastructure.
This is one huge commit to get this new version started.  During the
weekend, I played around different forms to re-implement ``tweedledum``.
The original idea to do it from zero but I soon discovered that (1) this
would take a lot of time and (2) I need a working version of the library
to do research!  So here I am on a sort of compromise:

I took most of the code from the previous version and change it into a
new directory structure and coding style.  I finally gave in to using
CamelCase for classes and structs, while still keeping snake_case for
all other names.  The reason is to not have code like ``angle angle``,
nor ``angle a_angle``.  Also, I am adopting the 80 column limit.

The idea is to do a redesign (or re-architect), while still being able
to develop new algorithms or modify existing ones.